Don't attempt to wrap each endpoint
Attempting to wrap every endpoint in the GitHub API with a python method is
bound to end in tears. Instead, just provide some simple get/put/post/etc.
helper wrappers around relative URLs and be done with it. This *barely*
increases the amount of code needed at point of use, and dramatically
simplifies the code API code.