Better support for porting pull requests
The migrator now attempts to label pull requests by ref first (in case the
branches are still around), falling back to labelling by SHA if that doesn't
work.

It will also add a comment to a merged pull request detailing when it was
merged, by whom (if it knows), and in which merge commit.